+++ packages/expo-updates/android/src/main/java/expo/modules/updates/UpdatesStateChangeSubscription.kt
@@ -1,0 +1,14 @@
+package expo.modules.updates
+
+import expo.modules.updatesinterface.UpdatesStateChangeSubscription
+
+class DisabledUpdatesStateChangeSubscription : UpdatesStateChangeSubscription {
+  override fun remove() {}
+}
+
+class EnabledUpdatesStateChangeSubscription(val subscriptionId: String) : UpdatesStateChangeSubscription {
+  override fun remove() {
+    val updatesController: EnabledUpdatesController? = UpdatesController.instance as EnabledUpdatesController
+    updatesController?.unsubscribeFromUpdatesStateChanges(subscriptionId)
+  }
+}